fix: fix E2E test verify Storage id
