+++ packages/tests/src/commonlib/constants.ts
@@ -149,8 +149,8 @@ export class EnvConstants {
   static readonly AAD_APP_OAUTH_AUTHORITY = "AAD_APP_OAUTH_AUTHORITY";
   static readonly AAD_APP_OAUTH_AUTHORITY_HOST = "AAD_APP_OAUTH_AUTHORITY_HOST";
   // FrontEnd
-  static readonly TAB_AZURE_STORAGE_RESOURCE_ID =
-    "TAB_AZURE_STORAGE_RESOURCE_ID";
+  static readonly AZURE_STATIC_WEB_APPS_RESOURCE_ID =
+    "AZURE_STATIC_WEB_APPS_RESOURCE_ID";
   static readonly TAB_AZURE_APP_SERVICE_RESOURCE_ID =
     "TAB_AZURE_APP_SERVICE_RESOURCE_ID";
   static readonly TAB_ENDPOINT = "TAB_ENDPOINT";
