+++ packages/tests/src/commonlib/index.ts
@@ -15,3 +15,4 @@ export * from "./appStudioValidator";
 export * from "./sharepointValidator";
 export * from "./staticSiteValidator";
 export * from "./containeraAppValidator";
+export * from "./utilities";
