+++ packages/tests/src/e2e/commonUtils.ts
@@ -13,12 +13,10 @@ import {
 } from "@microsoft/teamsfx-api";
 import { AzureScopes } from "@microsoft/teamsfx-core";
 import { dotenvUtil } from "@microsoft/teamsfx-core/src/component/utils/envUtil";
-import { exec } from "child_process";
 import * as dotenv from "dotenv";
-import fs from "fs-extra";
-import os from "os";
-import path from "path";
-import { promisify } from "util";
+import * as fs from "fs-extra";
+import * as os from "os";
+import * as path from "path";
 import { v4 as uuidv4 } from "uuid";
 import { YAMLMap, YAMLSeq, parseDocument } from "yaml";
 import {
@@ -30,6 +28,7 @@ import {
   ResourceGroupManager,
   SharepointValidator as SharepointManager,
   cfg,
+  getWebappServicePlan,
 } from "../commonlib";
 import {
   PluginId,
@@ -38,46 +37,15 @@ import {
   TestFilePath,
   fileEncoding,
 } from "../commonlib/constants";
-import { getWebappServicePlan } from "../commonlib/utilities";
 
 export const TEN_MEGA_BYTE = 1024 * 1024 * 10;
-export const execAsync = promisify(exec);
-
-export async function execAsyncWithRetry(
-  command: string,
-  options: {
-    cwd?: string;
-    env?: NodeJS.ProcessEnv;
-    timeout?: number;
-  },
-  retries = 3,
-  newCommand?: string
-): Promise<{
-  stdout: string;
-  stderr: string;
-}> {
-  const sleep = (ms: number) =>
-    new Promise((resolve) => setTimeout(resolve, ms));
-  while (retries > 0) {
-    retries--;
-    try {
-      const result = await execAsync(command, options);
-      return result;
-    } catch (e: any) {
-      console.log(
-        `Run \`${command}\` failed with error msg: ${JSON.stringify(e)}.`
-      );
-      if (e.killed && e.signal == "SIGTERM") {
-        console.log(`Command ${command} killed due to timeout`);
-      }
-      if (newCommand) {
-        command = newCommand;
-      }
-      await sleep(10000);
-    }
-  }
-  return execAsync(command, options);
-}
+export {
+  execAsync,
+  execAsyncWithRetry,
+  editDotEnvFile,
+  getProvisionParameterValueByKey,
+  getActivePluginsFromProjectSetting,
+} from "../commonlib";
 
 const testFolder = path.resolve(os.homedir(), "test-folder");
 
@@ -197,29 +165,6 @@ export async function setStaticWebAppSkuNameToStandardBicep(
 ): Promise<void> {
   const paramerters = { key: "staticWebAppSku", value: "Standard" };
   return setProvisionParameterValueV3(projectPath, envName, paramerters);
-}
-
-export async function getProvisionParameterValueByKey(
-  projectPath: string,
-  envName: string,
-  key: string
-): Promise<string | undefined> {
-  const parameters = await fs.readJSON(
-    path.join(
-      projectPath,
-      TestFilePath.configFolder,
-      `azure.parameters.${envName}.json`
-    )
-  );
-  if (
-    parameters.parameters &&
-    parameters.parameters.provisionParameters &&
-    parameters.parameters.provisionParameters.value &&
-    parameters.parameters.provisionParameters.value[key]
-  ) {
-    return parameters.parameters.provisionParameters.value[key];
-  }
-  return undefined;
 }
 
 export async function setBotSkuNameToB1(projectPath: string) {
@@ -524,21 +469,6 @@ export async function readContextMultiEnv(
   }
 }
 
-export async function getActivePluginsFromProjectSetting(
-  projectPath: string
-): Promise<any> {
-  const projectSettings = await fs.readJSON(
-    path.join(
-      projectPath,
-      TestFilePath.configFolder,
-      TestFilePath.projectSettingsFileName
-    )
-  );
-  return projectSettings[ProjectSettingKey.solutionSettings][
-    ProjectSettingKey.activeResourcePlugins
-  ];
-}
-
 export async function getCapabilitiesFromProjectSetting(
   projectPath: string
 ): Promise<any> {
@@ -781,32 +711,6 @@ export function getKeyVaultSecretReference(
   secretName: string
 ): string {
   return `@Microsoft.KeyVault(VaultName=${vaultName};SecretName=${secretName})`;
-}
-
-export function editDotEnvFile(
-  filePath: string,
-  key: string,
-  value: string
-): void {
-  try {
-    const envFileContent: string = fs.readFileSync(filePath, "utf-8");
-    const envVars: { [key: string]: string } = envFileContent
-      .split("\n")
-      .reduce((acc: { [key: string]: string }, line: string) => {
-        const [key, value] = line.split("=");
-        if (key && value) {
-          acc[key.trim()] = value.trim();
-        }
-        return acc;
-      }, {});
-    envVars[key] = value;
-    const newEnvFileContent: string = Object.entries(envVars)
-      .map(([key, value]) => `${key}=${value}`)
-      .join("\n");
-    fs.writeFileSync(filePath, newEnvFileContent);
-  } catch (error) {
-    console.log('Failed to edit ".env" file.');
-  }
 }
 
 export function removeTeamsAppExtendToM365(filePath: string) {
